refactor(ui): move FieldRenderer to @repo/ui and share across wizards
- FieldRenderer is pure form-field UI (renders a FieldDef); moved from
  apps/web/app/adopt/apply/_components to @repo/ui/components/field-
  renderer.tsx. Adopt and foster wizards now import the shared copy;
  the duplicated per-app files are removed.
- Reverts the /volunteer "Apply to Foster" CTA to mailto:fostering@
  gpa-mn.org to match the version that's been on main (the cherry-
  picked wizard commit brought the earlier link change back; keeping
  the mailto per prior direction).

Refs #12
